getAggregateQuery() and tests
